WAIFU-28: refactor: move user adapter to own module
